+++ package.json
@@ -100,8 +100,8 @@
     "start": "concurrently \"env-cmd -f ./.env.local react-scripts start\" \"wait-on http://localhost:3000 && env-cmd -f ./.env.local electron .\"",
     "premocha": "yarn prepare",
     "mocha": "concurrently \"env-cmd -f ./.env.test react-scripts start\" \"wait-on http://localhost:3000 && mocha --require @babel/register \"test/**/*.test.js\"\"",
-    "prettier:check": "prettier --check {src,public,test}/**/*.js",
-    "prettier:write": "prettier --write {src,public,test}/**/*.js",
+    "prettier:check": "prettier --check {src,public,test}/**{,/**}/*.js",
+    "prettier:write": "prettier --write {src,public,test}/**{,/**}/*.js",
     "hooks:uninstall": "node node_modules/husky/husky.js uninstall",
     "hooks:install": "node node_modules/husky/husky.js install",
     "ci": "react-scripts test --env=jsdom --coverage",
